% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF1.R
\name{ACF}
\alias{ACF}
\title{Autokovarianz-Funktion Schaetzer.}
\usage{
ACF(x, lag = NULL)
}
\arguments{
\item{x}{Eingabevektor der die beobachteten Daten enthaelt.}

\item{lag}{Zeitdifferenz der Datenreihe, welche die Funktion beruecksichtigen soll. Der Default-Wert ist die komplette Zeitreihe mit einem Wert weniger als die Anzahl der Beobachtungen der Zeitreihe.}
}
\value{
Der Rueckgabewert ist ein Vektor fuer die Abschaetzung der Autokorrelationsfunktion.
}
\description{
\code{ACF} berechnet eine Naeherung an die Autokovarianz-Funktion der Datenreihe X = x_1,...,x_n. Fuer mehr Informationen rufe die Vignette auf.
}
\examples{
 #Erstelle eine Zeitreihe
 X = arma_sim(phi=0.3,sd=1,I=1000)

 ACF(X, lag = 4)

}
