% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF1.R
\name{Autocovariance}
\alias{Autocovariance}
\title{Dies ist die Covarianz Funktion (Title).}
\usage{
Autocovariance(X, start = 1, end = length(X), type, h)
}
\arguments{
\item{X}{Eingabevektor}

\item{start}{Anfangszeitpunkt der Zeitreihe}

\item{end}{Endzeitpunkt der Zeitreihe}

\item{type}{Typ der Funktion ( Korrelation oder Kovarianz)}

\item{h}{Tutor fragen}
}
\value{
Der Rückgabewert ist eine Abschätzung der Autokorrelationsfunktion für eine observierte Datenreihe {X_t}.
}
\description{
\code{Autocovariance} berechnet die Covarianz des Eingabevektors x_1,...,x_n
}
\examples{
I <- 1000

X = arima.sim(n = I, list(ar = c(0.5, 0.499), ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796))

Autocovariance(X, lag.max = 4, type = "correlation", plot = FALSE)
}
