% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF1.R
\name{covariance}
\alias{covariance}
\title{Dies ist die Covarianz Funktion
Sie berechnet die Covarianz
Die ist manchmal groß
und manchmal klein}
\usage{
covariance(X, start = 1, end = length(X), h)
}
\description{
Dies ist die Covarianz Funktion
Sie berechnet die Covarianz
Die ist manchmal groß
und manchmal klein
}
