% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL_prediction.R
\name{DL_prediction}
\alias{DL_prediction}
\title{Durbin Levionson Prediction}
\usage{
DL_prediction(x, len = 1, all = FALSE)
}
\arguments{
\item{x}{Eingabevektor der die beobachteten Daten enth채lt}

\item{len}{Anzahl der vorherzusagenden Werten}

\item{all}{logical-Wert der angibt ob nur die vorhergesagten Werte oder die alte Zeitreihe plus die vorhergesagten Werte zur체ckgegeben wird. Der Default-Wert ist Ersteres.}
}
\value{
Vektor mit den vorhgesagten Werten
}
\description{
\code{DL_prediction} benutzt den Durbin-Levinson_Algorithmus (\code{DLA}) um eine Vorhersage f체r die Zeitreihe zu treffen. Mit dem \code{len} Parameter kann die Anzahl der zu vorhersagenden Werten getroffen werden. F체r mehr Informationen rufe die Vignette "Time-Series" auf.
}
\examples{
 #Erstelle eine Zeitreihe
 X = arma_sim(phi=0.3,sd=1,I=1000)

 DL_prediction(x,len=10)
}
