% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Durbin_Levinson_Algorithm.R
\name{DLA}
\alias{DLA}
\title{Durbin-Levinson Algorithm}
\usage{
DLA(x, len = NULL)
}
\arguments{
\item{x}{Eingabevektor, der die beobachteten Daten enthaelt.}

\item{len}{Horizont der Vorhersage.}
}
\value{
Vektor, welcher die Gewichtung der Phi's enthaelt.
}
\description{
\code{DLA} berechnet durch die rekursive Durbin-Levinson Methode die Gewichtung der Phi's fuer eine uebergebene Zeitreihe. Fuer mehr Informationen rufe die Vignette auf.
}
\examples{
 #Erstelle eine Zeitreihe
 X = arma_sim(phi=0.3,sd=1,I=1000)

 DLA(X, len = 4)
}
