% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{perio}
\alias{perio}
\title{Periodogramm}
\usage{
perio(y)
}
\arguments{
\item{y}{Eingabevektor, der die beobachteten Daten enthaelt.}
}
\value{
Vektor, der die Periodogrammwerte bzw. die Zeitreihe bei den Fourierfrequenzen enthaelt.
}
\description{
\code{perio} bestimmt die spektrale Dichte des Signals. Dadurch koennen die dominanten Frequenzen einer Zeitreihe identifiziert werden. Fuer mehr Informationen rufe die Vignette auf.
}
\examples{
 #Erstelle eine Zeitreihe
 X = arma_sim(phi=0.3,sd=1,I=1000)

 perio(X)
}
