% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_periodogram}
\alias{plot_periodogram}
\title{Periodogramm Plot}
\usage{
plot_periodogram(periodogram, logscale = F)
}
\arguments{
\item{periodogram}{Ein numerischer Vektor welcher ein Periodogram einer Zeitreihe enthält}

\item{logscale}{Wenn TRUE wird die Y-Achse logarithmisch skaliert.}
}
\value{
Plot des Periodograms in Abhängigkeit der Fourier Frequenzen
}
\description{
Diese Funktion plottet eine gegebenes PEriodogram.
}
\examples{
#Erstelle ein Periodogram einer Zeitreihe
X = perio(arma_sim(phi = c(0.5, 0.1), theta = c(-0.2), sd = 0.01, I = 100))
plot_periodogram(X)
}
