% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_timeseries1}
\alias{plot_timeseries1}
\title{Plot-Funktion}
\usage{
plot_timeseries1(timeseries)
}
\arguments{
\item{timeseries}{Ein numerischer Vektor vom Typ time series.}
}
\value{
Plot der Zeitreihe.
}
\description{
Diese Funktion plottet eine gegebene Timeseries (deutsch: Zeitreihe).
}
\examples{
#Erstelle eine Zeitreihe
X = arima.sim(n = 1000, list(ar = c(0.5, 0.499), ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796))
plot_timeseries(X)
}
