% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Innovation_Algorithm.R
\name{theta_sum}
\alias{theta_sum}
\title{Innovation Algorithm}
\usage{
theta_sum(theta, v, n, k)
}
\arguments{
\item{k}{}
}
\value{
Numerischer Vektor, der die Vorhersage des Innovation Algorithmus beinhaltet.
}
\description{
Beschreibung
Der Innovations Algorithmus bestimmt .

Details
}
\examples{
innovation(arima.sim(n = I, list(ar = c(0.8897,-0.4858),ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796)), 10)

}
