% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACF1.R
\name{ACF}
\alias{ACF}
\title{Autokovarianz-Funktion-Näherung.}
\usage{
ACF(x, lag = NULL)
}
\arguments{
\item{lag}{Zeitdifferenz der Datenreihe welche die Funktion berücksichtigen soll. Der Default Wert ist die komplette Zeitreihe mit einem Wert weniger als Observationen in der Zeitreihe}

\item{X}{Eingabevektor der die observierten Daten enthält}
}
\value{
Der Rückgabewert ist ein Vektor für die Abschätzung der Autokorrelationsfunktion.
}
\description{
\code{ACF} berechnet eine Näherung an die Autokovarianz-Funktion der Datenreihe X = x_1,...,x_n. Für mehr Informationen rufe die Vignette auf.
}
\examples{
 #Erstelle eine Zeitreihe
 X = arima.sim(n = 1000, list(ar = c(0.5, 0.499), ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796))

 ACF(X, lag = 4)

}
