% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMA_Generator.R
\name{arma_sim}
\alias{arma_sim}
\title{ARMA Generator}
\usage{
arma_sim(phi = 0, theta = 0, sd = 1, I)
}
\arguments{
\item{I}{}
}
\value{

}
\description{
Diese Funktion simuliert den ARMA-Prozess.

Das ARMA-Model beschreibt einen stochatischen Prozess mittels zwei Polynomen:
Der Autoregression (AR) und dem "moving average" (MA).
}
\examples{
gen()
gen()

}
