% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Durbin_Levinson_Algorithm.R
\name{DLA}
\alias{DLA}
\title{Durbin-Levinson Algorithm}
\usage{
DLA(x, len = NULL)
}
\arguments{
\item{x}{Eingabevektor der die beobachteten Daten enth채lt}

\item{len}{Horizont der Vorhersage.}
}
\value{
Vektor mit den vorhgesagten Werten
}
\description{
\code{DLA} berechnet durch die rekursive Durbin-Levinson Methode eine Vorhersage f체r die 체bergebene Zeitreihe. F체r mehr Informationen rufe die Vignette auf.
}
\examples{
 #Erstelle eine Zeitreihe
 X = arima.sim(n = 1000, list(ar = c(0.5, 0.499), ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796))

 DLA(X, len = 4)
}
