% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Innovation_Algorithm.R
\name{innovation}
\alias{innovation}
\title{Innovation Algorithm}
\usage{
innovation(ts, small_theta = matrix(0, lag, lag), lag = 1)
}
\arguments{
\item{ts}{Die Zeitreihe, welche uebergeben wird.}

\item{small_theta}{Optional eine Innovation-Matrix aus einer vorherigen Berechnung.}

\item{lag}{Der Index n bis zu welchem die Matrix berechnet werden soll.}
}
\value{
n x n Matrix mit den Innovation-Koeffizienten theta.
}
\description{
Der Innovation ist ein Vorhersage-Algorithmus, sagt fuer eine Zeitreihe also weitere Werte vorher.

Details
}
\examples{
innovation(arma_sim(phi = c(0.5, 0.1), theta = c(-0.2), sd = 0.01, I = 100))
}
