% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{perio}
\alias{perio}
\title{Periodogramm}
\usage{
perio(y)
}
\arguments{
\item{y}{Eingabevektor der die beobachteten Daten enthält}
}
\value{
Vektor der die Periodogrammwerte bzw. die Zeitreihe bei den Fourierfrequenzen enthält.
}
\description{
\code{perio} bestimmt die spektrale Dichte des Signals. Dadurch kann die dominaten Frequenzen einer Zeitreihe identifiziert werden. Für mehr Informationen rufe die Vignette auf.
}
\examples{
 #Erstelle eine Zeitreihe
 X = arima.sim(n = 1000, list(ar = c(0.5, 0.499), ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796))

 perio(X)
}
