% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_periodogram}
\alias{plot_periodogram}
\title{Plot-Funktion}
\usage{
plot_periodogram(periodogram)
}
\arguments{
\item{periodogram}{Ein numerischer Vektor welcher ein Periodogram einer Zeitreihe enthält}
}
\value{
Plot des Periodograms in Abhängigkeit der Fourier Frequenzen
}
\description{
Diese Funktion plottet eine gegebenes PEriodogram.
}
\examples{
#Erstelle ein Periodogram einer Zeitreihe
X = perio(arima.sim(n = 1000, list(ar = c(0.5, 0.499), ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796)))
plot_periodogram(X)
}
