% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plots}
\usage{
plot_timeseries(timeseries)
}
\arguments{
\item{timeseries}{is ein numerischer Vektor vom Typ time series.}
}
\value{
Grafische Ausgabe der Timeseries.
}
\description{
Diese Funktion plottet eine gegebene Timeseries (deutsch: Zeitreihe).
}
\examples{
plot_timeseries()
plot_timeseries()
}
