% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Time Series Plot}
\usage{
plot_timeseries(timeseries, pred = NULL, title = "Zeitreihe mit Vorhersage")
}
\arguments{
\item{timeseries}{Ein numerischer Vektor als Zeitreihe.}

\item{pred}{Optionaler Vektor einer Vorhersage der Zeitreihe}

\item{title}{Optional Character as Title}
}
\value{
Plot der Zeitreihe.
}
\description{
Diese Funktion plottet eine gegebene Timeseries (deutsch: Zeitreihe).
}
\details{
#'@import ggplot2
}
\examples{
#Erstelle eine Zeitreihe
X = arma_sim(phi = c(0.5, 0.1), theta = c(-0.2), sd = 0.01, I = 100)
f = ts_predict(X, 10)
plot_timeseries(X, f, title = "Vorhersage")
}
