% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction_Innovation.R
\name{ts_predict}
\alias{ts_predict}
\title{ts_predict}
\usage{
ts_predict(X, steps, all = FALSE)
}
\arguments{
\item{X}{Eine Stationäre Zeitreihe Zeitreihe}

\item{steps}{Anzahl N der Schritte der Prognose}

\item{all=FALSE}{}
}
\value{
Numerischer Vektor der Vorhersage. WEnn all=FALSE, werden nur die Vorhergesagten N Werte zurückgegeben, andernfalls der ganz Vektor X_hat
}
\description{
N-Schritt Prognose mithilfe des Innovation Algorithmus.

Details
}
\examples{
X = arima.sim(n = I, list(ar = c(0.8897,-0.4858),ma = c(-0.2279, 0.2488)), sd = sqrt(0.1796)), 10)
ts_predict(X, 4)

}
